backend/requirements.txt
# DO NOT  ADD: bcrypt or cffi explicitly !!
alembic==0.9.6
argh==0.26.1
aws-requests-auth==0.3.0
bleach==2.1.1
blinker==1.4
boto3==1.4.6
botocore==1.6.1
bugsnag==3.0.0
click==6.2
coaster==0.5.0
certifi==2016.9.26
docflow==0.3.2
editdistance==0.3.1
elasticsearch==5.3.0
enum34==1.1.2
factory-boy==2.8.1
Flask==0.12.0
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.2
Flask-Script==0.5.3
flask-socketio==2.9.2
Flask-SSLify==0.1.5
Flask-Testing==0.6.1
Flask-Compress==1.3.2
futures==3.0.5
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.4.5
gevent-websocket==0.10.1
html2text==2016.9.19
html5lib==1.0b10
hubstorage==0.22.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.5
MarkupSafe==1.0
matplotlib==2.1.0
newrelic==2.74.0.54
nose==1.3.7
ordereddict==1.1
pdfrw==0.3
psycopg2==2.7.1
py-bcrypt==0.4
pycparser==2.17
PyExecJS==1.1.0
Pygments==2.1.1
PyJWT==1.4.2
python-dateutil==2.6.1
python-dotenv==0.3.0
pytz==2015.7
PySocks==1.5.6
reportlab==3.3.0
requests==2.18.4
retrying==1.3.3
semantic-version==2.6.0
setuptools==36.7.2
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.11.0
slugify==0.0.1
SQLAlchemy==1.1.15
tldextract==1.7.5
UgliPyJS==0.2.5
Unidecode==0.4.19
urllib3==1.22
watchdog==0.8.3
webassets==0.12.0
WebOb==1.5.1
Werkzeug==0.11.4
wheel==0.26.0
tqdm==4.19.4
google-api-python-client==1.6.1
oauth2client==3.0.0
stripe==1.62.0
mandrill==1.0.57
mock==1.0.1
